requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8,<3.0
django-extensions>=3.0,<=4.0
black

gunicorn

djangorestframework
markdown
django-filter

django-json-widget
